docker/producer/requirements.txt
azure-eventhub
aioredis